Add macro to safely work with account lists
Accounts are a slice of AccountMeta / AccountInfo, and we need to match
them up by order. It is very easy to make a mistake in this if we do it
manually, and furthermore it is tedious to do all of the validation
properly. We can eliminate room for error by generating all of this with
a macro.