Add submodule for multisig repository
We want to use Serum's Multisig program for Solido governance initially.
We added a command-line interface for it, and to make it easy to manage
Solido, we want to extend that with Solido-specific options, so it makes
sense to move it into this repository. The on-chain program itself
remains unchanged though, so we can import the upstream repository as a
submodule here. There are a few patches needed for now, but I will
submit a pull request for those upstream, and I don't expect them to be
controversial.