Use account macros in tests
This is mostly a mechanical change to replace the instruction builder
function calls in the tests, to use the new account macros.

Unfortunately, due to this, the lifetime requirements of the entry point
have changed, and we need to patch Solana to be able to deal with that
in tests. For now we depend on our own fork that I patched.