Macro changes, minor changes in the order of accounts
I slightly changed the macro so we require signatures from the account
iff the account should be signed. In the same transaction we can execute
multiple instructions which will carry the same set of accounts, if some
of them are signed and some method does not require that, it will fail
where it shouldn't.