Make 'multisig' a subcommand of 'solido'
This unifies the two binaries, paving the way for proposing Solido
management operations to the governance multisig.

This does not yet update the tests. There are some other pending changes
to the tests to make them run in CI, I can rebase that on top of this
later and fix the command names at the same time.