Add Solana Program Library as a submodule
Solido depends on the SPL stake pool program, which is not present on a
local testnet by default, so we need to build it from source. To be able
to do that reliably in the tests, pull in the SPL as a submodule here,
so we can depend on its stake pool program.

The specific commit adds one patch on top of the upstream repository,
which removes the Cargo workspace in that repository, to prevent it from
interfering with our workspace.

Aside from helping to test on a localnet, having a pinned version of the
SPL stake pool program is also useful if we want to deploy our own
version of the stake pool: we can pin it, and upgrade at our own pace.
But we can still pull in the latest changes from upstream in the
submodule, there is no difficult patch backporting process.