Run build-bpf before test-bpf on CI
The test that depends on the SPL stake pool program fails without first
running "cargo build-bpf"; the fee computation is wrong. I'm not sure
why this happens, it's as if it uses an older version of the SPL stake
pool program, but I don't know where it gets that from. Either way,
running "cargo build-bpf" before running the tests makes the tests pass
...