Pass accounts to "solido create-solido" in test
This gets us one step further, but it is not yet sufficient: the
accounts are expected to be SPL token accounts, but to be able to
initialize them, we need to know the mint ... which only gets created
by create-solido.

I suppose we could have create-solido initialize these accounts, but I'm
not sure if the owners can still control them that way ... will
investigate further.