Break down setup transaction into multiple ones
The Solana transaction size limit is pretty aggressive, we can't set up
a Solido instance with all associated accounts in a single transaction,
it's much too big. So break it down into multiple transactions. This
means we lose atomicity, but that is not an essential property here. The
only downside is that if some transactions pass, but a later transaction
fails, we don't have the logic to resume, so you just spent some SOL
setting up rent-exempt accounts that you now have no way of reclaiming.
I don't think it's worth making the program deal with this though, setup
is a one-time thing.