Improve validator startup on CI
I think I figured out why the sleep was needed: the RPC can be available
before the test validator accepts transactions. Previously we only
waited for the RPC, but then subsequent operations could still fail,
because the validator was still initializing.

Fix this by not only waiting for the RPC to be available, but also
waiting until the validator has produced one block, which demonstrates
that it is validating, and therefore done initializing.