Typecheck and style-check Python code on CI
Typecheck is just useful to have, if you don't check it, then type
annotations will rot, and it will be ever more work to fix later.

Style check is more controversial, but right now we are ping ponging the
code between different formatters; we should settle on one, or not
autoformat at all. So let's settle on one.