Remove references to stake pool from program/src
This removes:

 * All references to stake pool accounts.
 * The stake pool related fields in the Lido struct.
 * Functions for dealing with stake pool tokens and computing balance.
 * The AccountType enum, it wasn't doing anything useful anyway.

As a drive-by, it also refactors the token mint check a bit, to make it
less verbose at the call site. What remains of the "check Lido for
deposit" check after deleting the stake pool, is just to check the SPL
token mint.

This commit makes the on-chain program and its tests compile again, but
the CLI program is broken now. I'll fix that in a follow-up commit,
because this one is already large.