Date metrics at the system time, not slot time
The slot time is wrong when you run a local validator. For one, it stops
ticking when you stop the validator, so you always have to start with a
new ledger. But even then, there seems to be a difference of a few
hours. They might have gotten UTC vs local time wrong? In any case,
let's use the SystemTime instead.