Take Pubkey by value in tests
This is a bit more convenient than by reference in some cases, when you
want to get the address from the Context, but also call a method on
Context. If you have to borrow &context.some_addr, then you cannot at
the same time call an &mut self method on Context (and most of them
are). But because Pukey is Copy, if we take it by value, then you can.