Set a safe minimum stake account balance
I just learned that there is an idiosyncrasy in Solana stake accounts,
that may prevent them from being merged. To avoid this footgun, we need
to make sure that a stake account contains sufficient balance, which is
supposedly more than the minimum rent-exempt balance.

Jon Cinque would check to see if it is possible to derive the minimum
amount from the credits_observed computations, but in the meantime,
1 SOL is supposed to be a safe amount.