Avoid non-rent-exempt accounts in tests
Ugh, I wasted half my morning on this ... transfer 100 lamports to an
account in epoch 95, and you will have a balance of 100 lamports. But
transfer 100 lamports to an account in epoch 20,000, and you will have
a balance of 0. This happens because rent needs to be paid up front, at
the first access to the account in the epoch. And apparently the rent in
the early epochs is lower?