Replace FeeDistribution with RewardDistribution
Previously when we used the stake pool, the stake pool would already be
configured with a fee cut, and we needed to split the fee according to
the fee distribution.

But now that we manage things ourselves, we get the full reward, and we
need to split it into:

 * A part that goes to stSOL value appreciation.
 * Fees. The fees we need to split up further.

To manage this, add one more field to the former FeeDistribution, so we
can control the share of stSOL value appreciation relative to other
fees. Then name then becomes inappropriate, so rename it as well.

Additionally, this has some small renames and refactors. The name
"distribute_fees" was a bit misleading, this function did not pay the
fees. So rename it to "split_reward", to clarify that it only computes
amounts, it does not pay anything.