Query accounts in multiple calls if needed
For now, I am testing with the devnet, and on the devnet public RPC
node, I can query at most 100 accounts per call ... so we need to break
up the calls. (Or spin up our own RPC node for the devnet, but that
would be even more work at this point.)

This gets the snapshot querying going, but unfortunately it still
creates torn reads. There is no way around that with multiple calls, but
we could create smarter batches, and group the reads in the order that
they happened, rather than hash map order. That would reduce the
probability of a tear being problematic.