Delay error when an account does not exist
Previously, we would load all accounts that we thought we would need,
and if one did not exist, that was a hard error. But, as we
opportunistically determine which accounts we *think* we will need, we
may try to load an account that does not exist, and fail on it, even
though we were never going to reference it.

This happend when I just started the daemon. In iteration 1, it merged
two stake accounts. Then in iteration 2, it failed, trying to load the
stake account at seed 0, which now no longer existed. We were not going
to reference it anyway, but we can only know that after the run
completes.

The solution is to not make missing accounts an error at snapshot time,
but only at the time when we try to access the account.