Update validator balance when withdrawing on merge
When we merge two stake accounts, we can withdraw inactive lamports back
to the reserve. But when we do this, we should update the tracked
validator balance, otherwise we double-count this SOL when computing the
total balance: it is now in the reserve, but we also still counted it as
part of the validator's stake accounts.

This adds a test that fails without the fix, and also implements the
fix.