Cache full ~/.cargo directory on CI
When you install a binary with "cargo install", it records some metadata
about the installed version in ~/.cargo/crates.toml, so a subsequent
"cargo install" can be a no-op. But if we cache the binary but not that
file, then "cargo install" will fail, complaining that it would
overwrite the file.