Add client functionality to withdraw.
Following the withdrawals implementation, this adds the functionality
for using the client to withdraw. The token is limited to the one
associated with the `signer`, just like in a deposit.
The function will delegate the tokens to the stake authority so they can
be burned.

Returns the newly created and delegated stake account, which will hold
the Lamports from the split.