Initialize Lido struct without mutation
This is mainly to make Clippy happy. It told me to not create a default
instance and then mutate it, but to use "Lido { fields,
..Lido::default() }" instead. But if we skip the mutation, we might as
well initialize the entire thing in one go. Only two fields had the
default value, so just spell that out, it’s more explicit anyway.