Specify mint and Solido keys.
Adds the possibility of specifying the minter and Solido keys for the
Solido initialization. Also creates the command `show-authorities` to
show the Solido authorities given a known public key.
When initializing the Solido instance with an existing key, the
initialization step changes slightly. Now it is also required to make
a new private key for the Solido instance and pass it in the
initialization method.

Adds extra check for the mint, now we require that the mint has the
Solido authority for minting and hasn't minted any tokens yet (supply is
zero).