Extend error printer to look up multisig errors
If we get "CustomError(105)", it is nice to now which error of the
multisig program (or Solido) that was, as the multisig program doesn't
log anything helpful. So try to convert back the error code into a typed
error, and print it.

Unfortunately, this does not seem to work for the error I am getting
currently (105, 0x69), which is returned from "multisig execute". Not
sure what is going on here, perhaps I am using the wrong conversion
method?