Cap withdrawal amounts to prevent imbalance
Thanks Vasily for this idea. We already force the user to withdraw from
the validator with the most balance, but users could previously still
drain that validator in a single large withdrawal, which would leave the
pool very unbalanced. By capping withdrawals at 10% of the validator's
stake account, if the pool is initially balanced, then this limits how
unbalanced a withdrawal can make the pool. Users who wish to withdraw
more can break up the withdrawal, and withdraw from multiple validators,
which then keeps the pool mostly balanced.