Improve type safety of metrics formatting
I happened to glance over this code, and then I realized that a boolean
field that says something about how to interpret a different field is
usually dangerous (because you can access the field directly without
checking the bool). With sum types, a better way is to make an enum with
an inner value.

So that's what I did here, remove the `nano` flag, and instead make an
enum that says whether the value is an integer, or a multiple of 1e-9.