Limit unstake accounts, fix amount computations
I wanted to put a limit on the number of unstake accounts, to make sure
that we don't have so many that WithdrawInactiveStake would fail,
because we don't merge the unstake accounts.

But while going over this, I noticed some funny things in the amount
computations. We should enforce the minimum stake account balance per
stake account, not for the validator's overall balance.

Also, when unstaking the full amount from an inactive validator, we
should unstake the full stake account, not the full amount, which might
be spread over multiple stake accounts. After a merge they should be in
a single one and then the difference is not relevant, but still, this
was a bit fishy, so fix that.