Do not perform maintenance unless we have duty
This is the main step towards making maitainer daemons take turns to
perform the maintenance: don't perform maintenance unless it's our turn.

There are still some things that can be improved. For example, if it is
not our turn, but it is almost our turn, then we shouldn't wait the
normal wait interval, we should instead wait until we expect it to be
our turn.