Add machinery to fetch validator info
This turned out to be quite a rabbit hole ... we need to load all
accounts owned by the Config Program, then find the ones that contain
validator info, and those accounts mention for which validator they
contain the info.

Then to make it interact with our snapshot reader, we have to identify
the config accounts that we are interested in in advance, and from then
on we know where we can get the validator info, so we can still do an
atomic read.