Preflight txs at confirmed level, wait for finalize
If we just set the RPC client commitment level to finalized, transaction
preflights would fail with this error:

  Solana RPC client returned an error:

    Request:    Some(SendTransaction)
    Kind:       RPC response error
    Error code: -32002
    Message:    Transaction simulation failed: Blockhash not found
    Reason:     Transaction preflight failure
    Error:

      Raw:      BlockhashNotFound
      Display:  Blockhash not found

Changing the commitment level for preflighting to confirmed resolves
this issue. This requires breaking up transaction sending into sending
and waiting for confirmation. (We need to wait for confirmation before
the changes done by our transaction are available in the next snapshot
we read, if we read at finalized commitment level.) This then uncovered
another Solana footgun, where if you want to wait for confirmation
without a spinner, you need to manually implement the wait loop.