Change some metrics from counter to gauge
On-chain these metrics really are counters, but because we read the
chain at "confirmed" commitment level and not "finalized", we might
observe a state that later does not get included in the chain, and in
that case, we can observe the value to go down.

At first I wanted to avoid this by reading only finalized chain states,
but it complicates other things (see previous commits), so instead we
will opt for making these metrics gauges instead of counters.