Use marker test address for Solido program too
This makes it a bit easier to identify programs in the "cargo test-bpf"
logs. Vanity addresses on mainnet are a bad idea because they seem like
they have some authority but they have zero authority at all and
teaching people to rely on this is dangerous. But for tests it's useful.