Bump CI cache version to cache Orca program
We download it into target/deploy, so it should be cached after one
download, so we don't hammer the Solana RPC servers for every CI run.
However, we need trigger a new cache, otherwise it will always download
the previous one that doesn’t have the file, and never upload the new
version.