Test that Anker::Deposit fails with wrong reserve
This is a bug that we had in the original Solido program. We added a
regression test for it there. We should add the same test here. Don't
make the same mistake twice. (The test fails, so we did not make the
mistake, but it is still good to confirm.)