Test Anker::Deposit fail with wrong instance addr
This depends on a feature that I am adding to the Solana program test
framework. For now I use a path override, which works for me locally,
but which will not work on CI. And Cargo Git dependencies can’t specify
a path as far as I am aware ... this is going to cause a bit of trouble.