Move error message out of exchange method
If we have the msg! unconditionally in there, then it will print every
time we generate an error, even if the error is handled properly. This
is a problem in the CLI program, because the exchange rate may be
undefined, and in that case we still need to output valid json, we can't
have arbitrary things being written to stdout.