Anker: Check which token swap program we call
This fixes a critical vulnerability identified by Neodyme: an attacker
could pass in their own program id for token_swap_program_id when
calling SellRewards, and we would happily invoke_signed that program,
which would allow it to steal our entire reserve.

The solution is to check that the presumed token swap program actually
owns the token swap instance. The token swap instance address is stored
in the Anker state and can only be changed by the manager, so it defines
what the correct swap program is.