Add additional balance check in Anker::SellRewards
The previous commit fixes a bug that would have enabled an attacker to
steal the full stSOL reserve. If this additional check would have been
in place (and the fix would not), then it wouldn't have prevented the
attacker from stealing some stSOL, but it would have prevented stealing
the full reserve: the attacker would only be able to steal the excess
that we planned to sell.