Pin Click version used on CI
The lint check is failing with some error related to click. It looks
like in the install phase it tries to uninstall some existing
installation of 7.0 while also installing 8.1 ... that looks like it
might be a problem, I am going to try to pin this.

...

We should just have used Nix already and pinned everything ...