Add test for Anker::SendRewards stack overflow
We cannot really test this call without the full Wormhole programs and
all of their associated accounts and dependencies in place, which is a
pain. But we can at least test that can get one particular error. I am
hoping that failing to call the Wormhole program will result in a
different error than the "ProgramFailedToComplete" that we get now, and
that we can test for the stack overflow this way.

The new test currently fails.